Correct OnMoved handler for StackNSplit.
Fixed issue causing items moved to other stackable (of the same type) or
item container to be granted charges instead of the item (swapped) that
said item was being stacked with.
This issue was hidden due to stackables sharing the same type but is
visible when container item is engaged in the inventory manipulation
event.

Item charges manipulation and event firing has been corrected to match adequate
items during such exchange.