Account for dependencies when compiling
Update build script so that all dependencies are added to the source
file list before compilation starts. For project to compile, ensure that
the map initialization functions found in 'war3map.j' are located at the
bottom of the generated file.

As script is intended for compilation of jass files found within 'src'
directory, drop JassSrc variable and declare new members to denote
project's specifics.